Moving studio logic to its own file, and adding an unused timestamp field to trigger file watcher
